fix: tons como decimal en PlanMonthDto

+++ haulages_bot/Controllers/ProductionPlanBotController.cs
@@ -412,7 +412,7 @@ namespace haulages_bot.Controllers
             public int Month { get; set; }
 
             [JsonProperty("tons")]
-            public int Tons { get; set; }
+            public decimal Tons { get; set; }
 
             [JsonProperty("minerals")]
             public List<PlanMineralDto>? Minerals { get; set; }
